Rename BlockDownloader to BlockDownloaderOld
