+++ packages/p2p/source/service-provider.ts
@@ -3,7 +3,7 @@ import { Providers, Services, Utils } from "@mainsail/kernel";
 import Joi from "joi";
 
 import { ValidateAndAcceptPeerAction } from "./actions";
-import { BlockDownloader } from "./block-downloader";
+import { BlockDownloader } from "./block-downloader-old";
 import { Broadcaster } from "./broadcaster";
 import { ChunkCache } from "./chunk-cache";
 import { Downloader } from "./downloader";
